Split out littlefs mbed-os implementation from 'ca7b5e8eed574a7295280d9276de3f9bb7cb7f05'
This is not a true subtree split due to the complication moves in the
mbed-os tree. While I could patch the history together to keep the
subtree intact, the ultimate goal for this split is to be merged in
parallel to littlefs-v1. This would give us duplicate histories anyway
so it's probably better to keep this split as a single squashed commit.

source-dir: features/storage/filesystem/littlefs/littlefs
source-hash: ca7b5e8eed574a7295280d9276de3f9bb7cb7f05